(#145) Cria componente filho para listagem projetos
