+++ puma/src/components/projeto/consulta-projeto/ListagemConsultaProjetoComponent.vue
@@ -1,0 +1,8 @@
+<template>
+  <div>
+    <mdbDatatable ref="table" borderless color="blue" :data="data" :pagination="false" />
+    <div v-if="!projects.length" class="no-results align-content-center mt-3">
+      Sem resultados disponíveis
+    </div>
+  </div>
+</template>
