--- puma/src/components/projeto/consulta-projeto/consulta-projeto.html
@@ -1,13 +1,0 @@
-<div class="main-content hidden-scroll-bar">
-  <div class="tittle ml-4">
-    {{ operacao === 'meus-projetos' ? 'Meus projetos' : (operacao === 'projetos' ? 'Projetos' : 'Projetos das Disciplinas') }}
-  </div>
-  <div v-if="projects.length" class="tittle sub-tittle ml-4">
-    {{ 'Projetos ' + (operacao === 'meus-projetos' ? 'cadastrados pelo usuário' : (operacao === 'projetos' ?
-    'cadastrados na plataforma' : 'alocados para as disciplinas ministradas pelo usuário')) }}
-  </div>
-  <mdbDatatable ref="table" borderless color="blue" :data="data" :pagination="false" />
-  <div v-if="!projects.length" class="no-results align-content-center mt-3">
-    Sem resultados disponíveis
-  </div>
-</div>
